todo/serializers.py
from rest_framework import serializers

from todo.models import TodoList, Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ('id', 'name',)


class TodoListSerializer(serializers.ModelSerializer):
    class Meta:
        model = TodoList
        fields = ('id', 'title', 'content', 'created', 'due_date', 'completed')
